requirements.txt
psycopg2
Django
djangorestframework
reportlab
django-filter
drf_yasg==1.20.0